+++ rho_sum_calc.cu
@@ -19,9 +19,9 @@
 #include <thrust/host_vector.h>
 #include <thrust/device_vector.h>
 
-extern "C" float get_signal_gpu(GPU_data *gpu_setup, int L, int R, int n, int num_blocks, int num_threads);
+extern "C" double get_signal_gpu(GPU_data *gpu_setup, int L, int R, int n, int num_blocks, int num_threads);
 
-__global__ void cal_esum1(int L, int R, float *rho_sum, float *rho_e, float *w_pot, int max_threads){
+__global__ void cal_esum1(int L, int R, double *rho_sum, float *rho_e, float *w_pot, int max_threads){
     // esum1 += rho_e[0][z][r] * (double) (r-1) * setup.wpot[r-1][z-1];
     int r = blockIdx.x%R;
     int z = (floorf(blockIdx.x/R) * max_threads) + threadIdx.x;
@@ -29,10 +29,10 @@ __global__ void cal_esum1(int L, int R, float *rho_sum, float *rho_e, float *w_p
     if(r==0 || z==0 || r>=R || z>=L){
       return;
     }
-  rho_sum[((R+1)*z)+r] = rho_e[(0*(L+1)*(R+1))+((R+1)*z)+r] * (float) (r-1) * w_pot[((R+1)*(z-1))+(r-1)];
+  rho_sum[((R+1)*z)+r] = rho_e[(0*(L+1)*(R+1))+((R+1)*z)+r] * (double) (r-1) * w_pot[((R+1)*(z-1))+(r-1)];
 }
 
-__global__ void cal_esum2(int L, int R, float *rho_sum, float *rho_e, int max_threads){
+__global__ void cal_esum2(int L, int R, double *rho_sum, float *rho_e, int max_threads){
   // esum2 += rho_e[0][z][r] * (float) (r-1);
   int r = blockIdx.x%R;
   int z = (floorf(blockIdx.x/R) * max_threads) + threadIdx.x;
@@ -40,10 +40,10 @@ __global__ void cal_esum2(int L, int R, float *rho_sum, float *rho_e, int max_th
   if(r==0 || z==0 || r>=R || z>=L){
     return;
   }
-  rho_sum[((R+1)*z)+r] = rho_e[(0*(L+1)*(R+1))+((R+1)*z)+r] * (float) (r-1);
+  rho_sum[((R+1)*z)+r] = rho_e[(0*(L+1)*(R+1))+((R+1)*z)+r] * (double) (r-1);
 }
 
-__global__ void cal_hsum1(int L, int R, float *rho_sum, float *rho_h, float *w_pot, int max_threads){
+__global__ void cal_hsum1(int L, int R, double *rho_sum, float *rho_h, float *w_pot, int max_threads){
   // hsum1 += rho_h[0][z][r] * (float) (r-1) * setup.wpot[r-1][z-1];
   int r = blockIdx.x%R;
   int z = (floorf(blockIdx.x/R) * max_threads) + threadIdx.x;
@@ -51,10 +51,10 @@ __global__ void cal_hsum1(int L, int R, float *rho_sum, float *rho_h, float *w_p
   if(r==0 || z==0 || r>=R || z>=L){
     return;
   }
-  rho_sum[((R+1)*z)+r] = rho_h[(0*(L+1)*(R+1))+((R+1)*z)+r] * (float) (r-1) * w_pot[((R+1)*(z-1))+(r-1)];
+  rho_sum[((R+1)*z)+r] = rho_h[(0*(L+1)*(R+1))+((R+1)*z)+r] * (double) (r-1) * w_pot[((R+1)*(z-1))+(r-1)];
 }
 
-__global__ void cal_hsum2(int L, int R, float *rho_sum, float *rho_h, int max_threads){
+__global__ void cal_hsum2(int L, int R, double *rho_sum, float *rho_h, int max_threads){
   // hsum2 += rho_h[0][z][r] * (float) (r-1);
   int r = blockIdx.x%R;
   int z = (floorf(blockIdx.x/R) * max_threads) + threadIdx.x;
@@ -62,10 +62,10 @@ __global__ void cal_hsum2(int L, int R, float *rho_sum, float *rho_h, int max_th
   if(r==0 || z==0 || r>=R || z>=L){
     return;
   }
-  rho_sum[((R+1)*z)+r] = rho_h[(0*(L+1)*(R+1))+((R+1)*z)+r] * (float) (r-1);
+  rho_sum[((R+1)*z)+r] = rho_h[(0*(L+1)*(R+1))+((R+1)*z)+r] * (double) (r-1);
 }
 
-extern "C" float get_signal_gpu(GPU_data *gpu_setup, int L, int R, int n, int num_blocks, int num_threads){
+extern "C" double get_signal_gpu(GPU_data *gpu_setup, int L, int R, int n, int num_blocks, int num_threads){
 
     double signal, hsum1=0.f, hsum2=0.f, esum1=0.f, esum2=0.f;
     // esum1 += rho_e[0][z][r] * (float) (r-1) * setup.wpot[r-1][z-1];
